Move IdProvider impls to own module
