+++ ws-server/src/server.rs
@@ -37,11 +37,12 @@ use futures_channel::mpsc;
 use futures_util::future::{join_all, FutureExt};
 use futures_util::io::{BufReader, BufWriter};
 use futures_util::stream::StreamExt;
+use jsonrpsee_core::id_providers::RandomIntegerIdProvider;
 use jsonrpsee_core::middleware::Middleware;
 use jsonrpsee_core::server::helpers::{collect_batch_response, prepare_error, MethodSink};
 use jsonrpsee_core::server::resource_limiting::Resources;
 use jsonrpsee_core::server::rpc_module::{ConnectionId, MethodResult, Methods};
-use jsonrpsee_core::traits::{IdProvider, RandomIntegerIdProvider};
+use jsonrpsee_core::traits::IdProvider;
 use jsonrpsee_core::{Error, TEN_MB_SIZE_BYTES};
 use soketto::connection::Error as SokettoError;
 use soketto::handshake::{server::Response, Server as SokettoServer};
@@ -695,7 +696,8 @@ impl<M> Builder<M> {
 	///
 	/// ```rust
 	/// use jsonrpsee_ws_server::WsServerBuilder;
-	/// use jsonrpsee_core::traits::{IdProvider, RandomStringIdProvider};
+	/// use jsonrpsee_core::traits::IdProvider;
+	/// use jsonrpsee_core::id_providers::RandomStringIdProvider;
 	///
 	/// let builder = WsServerBuilder::default().set_id_provider(RandomStringIdProvider::new(16));
 	/// ```
